+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.0 (unreleased)
+
+FIXES
+
+* Email addresses are comparsed case insensitively when determining if a user
+  needs to be replaced.
+
 ## 0.1.0 (11 July 2022)
 
 FEATURES: